Rename internal/net to internal/netutils
Avoid future collisions with `net` standard library package.

refs GH-140
