+++ cmd/lscert/main.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/atc0005/check-certs/internal/certs"
 	"github.com/atc0005/check-certs/internal/config"
-	"github.com/atc0005/check-certs/internal/net"
+	"github.com/atc0005/check-certs/internal/netutils"
 	"github.com/atc0005/check-certs/internal/textutils"
 	"github.com/atc0005/go-nagios"
 )
@@ -81,7 +81,7 @@ func main() {
 	case cfg.Server != "":
 
 		var certFetchErr error
-		certChain, certFetchErr = net.GetCerts(cfg.Server, cfg.Port, cfg.Timeout(), log)
+		certChain, certFetchErr = netutils.GetCerts(cfg.Server, cfg.Port, cfg.Timeout(), log)
 		if certFetchErr != nil {
 			log.Error().Err(certFetchErr).Msg(
 				"error fetching certificates chain")
