+++ internal/netutils/doc.go
@@ -5,6 +5,6 @@
 // Licensed under the MIT License. See LICENSE file in the project root for
 // full license information.
 
-// Package net provides helper functions for network related operations such
-// as port scanning or subnet slicing.
-package net
+// Package netutils provides helper functions for network related operations
+// such as port scanning or subnet slicing.
+package netutils
