+++ internal/netutils/net.go
@@ -5,7 +5,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for
 // full license information.
 
-package net
+package netutils
 
 import (
 	"crypto/tls"
